fix: move the implementation of OAuth state generation to the backend
Remove the anchor tag from the OAuth sign in button and send a request
to backend to generate the OAuth state. Send the OAuth state as a
cookie and the OAuth URL as the response body to the client. Make the
cookie property 'HttpOnly' true which was false when the cookie was
created in frontend. Redirect the user to the OAuth URL using
window.location.href = oauthUrl instead of using an anchor tag.