rm makefile, nginx. Add caddy.
It now uses Caddy as webserver which is way simpler than nginx.
Caddy can generate certificates itself so we don't have to handle them.
The frontend docker compose config now just starts once, builds the files and then exits again. This way caddy has the new static files to serve. There is no need to keep the frontend running itself as its just static files?
Also don't expose the backend port in the docker config as it is only required inside the docker network.
Doesn't use https for the backend anymore as internally ssl is not needed and just uneccessary complexity. SSL is fully handled by caddy which is the only endpoint we intend to expose to the world.